fix: follow_symlinks is commented out
When removing a folder, any `follow_symlinks` options are commented out
by Sublime. Because Sublime keeps comments in place, adding and removing
the same project many times leads to a build up of comments in the
projects file.

As it seems the follow_symlinks option is not used by Sublime, not
adding it when adding folders resolves this issue.